Disable Resumable Uploads
In some cases, resumable uploads cannot happen on GitHub Actions because
the `~/.config` directory is not writable on their Windows instances.
This leads to an error that cannot be solved in the project alone. To
get around this, add the ability for users of this package to configure
whether they want resumable uploads or not. The `resumable` setting can
be configured in the Forge config for this publisher.

Fixes #7